organization controller improved

+++ src/main/java/com/example/authservice/controller/OrganizationController.java
@@ -1,9 +1,11 @@
 package com.example.authservice.controller;
 
 
+import com.example.authservice.dto.ChangePasswordDto;
 import com.example.authservice.dto.OrganizationDto;
 import com.example.authservice.model.Organization;
 import com.example.authservice.service.OrganizationService;
+import com.example.authservice.util.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -18,17 +20,35 @@ import java.util.Objects;
 @RequestMapping("/organization")
 public class OrganizationController {
 
-    @Autowired
+    final
     OrganizationService organizationService;
 
+    public OrganizationController(OrganizationService organizationService) {
+        this.organizationService = organizationService;
+    }
+
     @PostMapping("/register")
+    //completed
     public Organization registerOrganization(@RequestBody OrganizationDto organizationDto){
         return organizationService.registerOrganization(organizationDto);
     }
 
     @GetMapping("/{id}")
-    public Organization getOrganizaion(@PathVariable Long id){
-        return organizationService.getOrganizationById(id);
+    public ResponseEntity<?> getOrganizaion(@PathVariable Long id,@AuthenticationPrincipal UserDetails userDetails){
+    //completed
+
+        if(SecurityUtils.isAdmin(userDetails.getAuthorities())){
+            return ResponseEntity.status(HttpStatus.OK).body(organizationService.getOrganizationById(id));
+        }
+
+        if(SecurityUtils.isOrganization(userDetails.getAuthorities())){
+            Organization organization = organizationService.getOrganizationByEmail(userDetails.getUsername());
+            if(Objects.equals(id, organization.getId())|| SecurityUtils.isAdmin(userDetails.getAuthorities())){
+                return ResponseEntity.status(HttpStatus.OK).body(organizationService.getOrganizationById(id));
+            }
+        }
+
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized: you are not the account user.");
     }
 
     @PutMapping("/{id}")
@@ -37,12 +57,18 @@ public class OrganizationController {
             @RequestBody OrganizationDto organizationDto,
             @AuthenticationPrincipal UserDetails userDetails
     ){
-        System.out.println(userDetails.getAuthorities());
 
-        Organization existingOrganization = organizationService.getOrganizationByEmail(userDetails.getUsername());
+        //completed
 
-        if(Objects.equals(existingOrganization.getId(), id)){
+        if (SecurityUtils.isAdmin(userDetails.getAuthorities())){
+
             return ResponseEntity.status(HttpStatus.OK).body(organizationService.updateOrganization(id,organizationDto));
+
+        }
+
+        if(SecurityUtils.isOrganization(userDetails.getAuthorities())){
+            Organization existingOrganization = organizationService.getOrganizationByEmail(userDetails.getUsername());
+            return ResponseEntity.status(HttpStatus.OK).body(organizationService.updateOrganization(existingOrganization.getId(),organizationDto));
         }
 
         return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized: you are not the account user.");
@@ -50,45 +76,57 @@ public class OrganizationController {
     }
 
     @DeleteMapping("/{id}")
-    public String deleteOrganization(@PathVariable Long id,@AuthenticationPrincipal UserDetails userDetails){
+    public ResponseEntity<?> deleteOrganization(@PathVariable Long id,@AuthenticationPrincipal UserDetails userDetails){
 
-        organizationService.deleteOrganization(id);
-        return "organization has been deleted";
+        if (SecurityUtils.isAdmin(userDetails.getAuthorities())){
+            organizationService.deleteOrganization(id);
+            ResponseEntity.status(HttpStatus.OK).body("user has been deleted");
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("only admin can delete the accout ");
     }
 
     @GetMapping("/")
-    public List<Organization> getAllOrganizations(){
-        return organizationService.findAllOrganizations();
+    public ResponseEntity<?> getAllOrganizations(@AuthenticationPrincipal UserDetails userDetails){
+         //only for admin
+        if (SecurityUtils.isAdmin(userDetails.getAuthorities())){
+            return ResponseEntity.status(HttpStatus.OK).body(organizationService.findAllOrganizations());
+        }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("only admin can access all organizations");
     }
 
     @PatchMapping("/changepassword")
-    public ResponseEntity<?> changePassword(@RequestBody OrganizationDto organizationDto, @AuthenticationPrincipal UserDetails userDetails) {
+    public ResponseEntity<?> changePassword(@RequestBody ChangePasswordDto changePasswordDto, @AuthenticationPrincipal UserDetails userDetails) {
+        if (SecurityUtils.isOrganization(userDetails.getAuthorities())){
+            Organization organization = organizationService.getOrganizationByEmail(userDetails.getUsername());
+            organizationService.changeOrganizationPassword(organization.getId(),changePasswordDto.getOldPassword(),changePasswordDto.getNewPassword());
+            return ResponseEntity.status(HttpStatus.OK).body("password has been changedd");
 
-        Organization organization = organizationService.getOrganizationByEmail(userDetails.getUsername());
-        if (Objects.equals(organization.getId(), organizationDto.getId())) {
-            organizationService.changeOrganizationPassword(
-                    organizationDto.getId(),
-                    organizationDto.getPassword()
-            );
-            return ResponseEntity.ok("The password has been changed.");
-        } else {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized: you are not the account user.");
         }
+
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("only the owner of the user can change the password");
     }
 
     @PatchMapping("/activite/{id}")
-    public String activateAccount(@PathVariable Long id){
+    public ResponseEntity<?> activateAccount(@PathVariable Long id,@AuthenticationPrincipal UserDetails userDetails){
 
-        organizationService.activateOrganizationAccount(id);
-        return "account has been activated";
+        if (SecurityUtils.isAdmin(userDetails.getAuthorities())){
+            organizationService.activateOrganizationAccount(id);
+            return ResponseEntity.status(HttpStatus.OK).body("account has been activited");
+        }
+
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("account not authorized");
 
     }
 
     @PatchMapping("/deactivate/{id}")
-    public String deactivateAccount(@PathVariable Long id){
-        organizationService.deactivateOrganizationAccount(id);
-        return "account has been deactivated";
-    }
+    public ResponseEntity<?> deactivateAccount(@PathVariable Long id,@AuthenticationPrincipal UserDetails userDetails){
 
+        if (SecurityUtils.isAdmin(userDetails.getAuthorities())){
+            organizationService.deactivateOrganizationAccount(id);
+            return ResponseEntity.status(HttpStatus.OK).body("account has been deactivated");
+        }
+
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("only admin can effect this operation");
+    }
 
 }
